LayoutTests/media/encrypted-media/mock-MediaKeySystemAccess.html

<!DOCTYPE html>
<html>
<head>
    <script src=../video-test.js></script>
    <script type="text/javascript">
    var mock;
    var promise;
    var mediaKeySystemAccess;
    var capabilities = {};

    function doTest()
    {
        if (!window.internals) {
            // failTest("Internals is required for this test.")
            // return;
        }

        run('internals.initializeMockMediaSource()');
        run('mock = internals.registerMockCDM()');
        run('mock.supportedDataTypes = ["mock"]');

        next();
    }

    function next() {
        if (!tests.length) {
            mock.unregister();
            endTest()
            return;
        }

        var nextTest = tests.shift();
        consoleWrite('');
        nextTest();
    }

    function shouldResolve(promise) {
        promise.then(mediaKeySystemAccess => {
            logResult(Success, 'Promise resolved');
            next();
        }, () => {
            logResult(Failed, 'Promise rejected');
            next();
        });
    }

    function shouldReject(promise) {
        promise.then(() => {
            logResult(Failed, 'Promise resolved incorrectly');
            next();
        }, exceptionCode => {
            logResult(Success, 'Promise rejected correctly');
            next();
        });
    }

    function gotMediaKeySystemAccess(result) {
        logResult(Success, 'Promise resolved');
        mediaKeySystemAccess = result;
        next();
    }

    tests = [
        function() {
            run('capabilities.initDataTypes = ["mock"]');
            run(`capabilities.videoCapabilities = [{ contentType: 'video/mock; codecs="mock"' }] `);
            run('promise = navigator.requestMediaKeySystemAccess("org.webkit.mock", [capabilities])');
            promise.then(gotMediaKeySystemAccess).catch(failTest);
        },

        function() {
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldResolve(promise);
        },

        function() {
            run('mock.canCreateInstances = false');
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldReject(promise);
        },

        function() {
            run('mock.canCreateInstances = true');
            run('capabilities.distinctiveIdentifier = "not-allowed"');
            run('promise = navigator.requestMediaKeySystemAccess("org.webkit.mock", [capabilities])');
            promise.then(gotMediaKeySystemAccess).catch(failTest);
        },

        function() {
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldResolve(promise);
        },

        function() {
            run('mock.distinctiveIdentifiersRequirement = "required"');
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldReject(promise);
        },

        function() {
            run('mock.distinctiveIdentifiersRequirement = "optional"');
            run('capabilities.persistentState = "not-allowed"');
            run('promise = navigator.requestMediaKeySystemAccess("org.webkit.mock", [capabilities])');
            promise.then(gotMediaKeySystemAccess).catch(failTest);
        },

        function() {
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldResolve(promise);
        },

        function() {
            run('mock.persistentStateRequirement = "required"');
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldReject(promise);
        },
    ];
    </script>
</head>
<body onload="doTest()">
</body>
</html>
